refactor axios requests and render behaviour of the UserlocationGallery upon pressing the delete button
